LPD-50859 Move package
